+++ src/utilities/lineRemover.js
@@ -30,12 +30,9 @@ class lineRemover {
      */
     removeSingleLineComments() {
         for (let i = 0; i < this.lineContent.length; i++) {
-            let hexadecimal = this.lineContent[i].replace(/^((?!:).)*\/\/.*/g, '');
-            if (hexadecimal != null) {
-                this.lineContent[i].replace(/^((?!:).)*\/\/.*/g, '');
-                this.lineContent[i] = hexadecimal;
-            }
+            this.lineContent[i] = this.lineContent[i].replace(/\/\/.*/g, '');
         }
+
     }
 
     /**
@@ -47,19 +44,19 @@ class lineRemover {
      */
     removeMultipleLineComments() {
         for (let i = 0; i < this.lineContent.length; i++) {
-            let begin = this.lineContent[i].match(/(\/\*)/ig); // first /* found
+            let begin = this.lineContent[i].match(/(\/\*)/gm); // first /* found
             if (begin != null) {
                 for (let j = 0; j < this.lineContent.length; j++) {
-                    let end = this.lineContent[j].match(/(\*\/)/g); //found */ end
+                    let end = this.lineContent[j].match(/(\*\/)/gm); //found */ end
                     if (end != null) {
                         for (let k = i; k < j + 1; k++) {
                             if (k == i) {
-                                let FirstCharacterToRemove = this.lineContent[k].indexOf("/*");
+                                let FirstCharacterToRemove = this.lineContent[k].indexOf("/");
                                 let firstLineToReplace = this.lineContent[k].substring(0, FirstCharacterToRemove);
                                 this.lineContent[k] = firstLineToReplace;
                             } else if (k == j) {
                                 let lastCharacterToRemove = this.lineContent[k].indexOf("*/");
-                                let lastLineToReplace = this.lineContent[k].substring(lastCharacterToRemove + 2, this.lineContent[k].length);
+                                let lastLineToReplace = this.lineContent[k].substring(lastCharacterToRemove + 2, this.lineContent[k].length + 1);
                                 this.lineContent[k] = lastLineToReplace;
 
                             } else {